Moderate progress on #120. Details below.
The editor and its related editorProgram component are currently designed
to let a user save bits of code that run in the context of an existing
level. This is great for training with interactive commands against worlds,
but it's terrible for authoring a brand new world.

It's possible to make this component polymorphic, but it takes work.

The current spike takes the originating levelId from the base program context,
and saves it on the editorProgram template.data.

To make this fully functional:

* Allow specifying an editor mode for program or world
* When in world mode, it will attempt to see if a world exists for the
current user with a (too be added property) called baseLevelId or
levelOriginId or something like that. If it does not exist, it will
save the world for the current user and then render the game component
with that newly created levelId. From this point, the Update button will
save the entire script over the world like a normal build.js page does. It
will then re-render the game component to update that view. Challenges may
be timers and static/globalish state in the game component left over from
ages ago. As a workaround, we could make it pop up the newly saved world
in a popup window just we do on build.js -- or we can render a very slimmed
down version of play.js inside an iframe to avoid those timer & globalish
issues.